--- a/wave_viewer.m
+++ b/wave_viewer.m
@@ -1688,11 +1688,7 @@ else
         if ~isfield(marker_name_spec,'iidatsource4ypos'), marker_name_spec.iidatsource4ypos = []; end
         marker_name_spec.h_tmarker = h_tmarker;
         marker_name_str = get_marker_name_str(marker_name_spec,t);
-        cur_hax = gca;
- 
-%          axes(hax);  % because the text() command only works with the current axes 
-        h_marker_name(i_marker_name) = text(hax,t,tmarker_ydat(2),marker_name_str); % text(hax,...) updates axes automatically
-%         axes(cur_hax);
+        h_marker_name(i_marker_name) = text(hax,t,tmarker_ydat(2),marker_name_str);
 
         set(h_marker_name(i_marker_name),'VerticalAlignment',marker_name_spec.vert_align);
         if strcmp(marker_name_spec.color,'same')