--- a/wave_viewer.m
+++ b/wave_viewer.m
@@ -1689,7 +1689,6 @@ else
         marker_name_spec.h_tmarker = h_tmarker;
         marker_name_str = get_marker_name_str(marker_name_spec,t);
         h_marker_name(i_marker_name) = text(hax,t,tmarker_ydat(2),marker_name_str);
-
         set(h_marker_name(i_marker_name),'VerticalAlignment',marker_name_spec.vert_align);
         if strcmp(marker_name_spec.color,'same')
             set(h_marker_name(i_marker_name),'Color',get(h_tmarker,'Color'));
@@ -1703,7 +1702,6 @@ end
 set(h_tmarker,'UserData',h_marker_name);
 if ~yes_visible, set(h_tmarker,'Visible','off'); end
 end
-
 
 function update_tmarker(h_tmarker,t)
 if isempty(t)