To improve the load time for user events. Previously, a separate call to cur_hax (the previous axis information) was required to restore properties of user events.

What we tried:
Separate function call: axes creation within for-loop was slow, so we moved it to another function and called it from make_tmarker
Parallel programming: create multiple 'workers' to perform tasks simultaneously (create all the user events at once)

Final changes:
Add the 'hax' parameter to the text function call. 

Edge cases tested:
Multiple overlapping events
Deleting multiple events and reloading audioGUI
Moving user events and expanding
Changing pre-emphasis or LPC order

Running profiler, we observe the following for make_tmarker:

For a trial with 3 user events, an ~87% decrease in runtime is observed.

Test 1:
Before: 1.086 seconds
After: 0.144 seconds

Test 2:
Before: 1.085 seconds
After: 0.148 seconds

etc.


For a trial with 20 user events, a ~95% decrease in runtime is observed.
Before: 4.947 seconds
After: 0.210 seconds

etc.