@cwnaber 2024-11-06
@carrien , I think this change is ready to go, but wanted to send it by your desk first, mostly as an FYI.

Another brief description of the issue/solution: wave_viewer uses the MATLAB built-in function `text` to create the label attached to user events, ie, the object which displays "uev1" next to a user event. Previously, `text` could only create a text object in the current axis, so wave_viewer would switch axis, make the text object, then switch back to the previous axis to create a label for each axis. Switching the current axis is very slow. Now, you can specify an axis when you call `text`, so no switching of the current axis is needed.